requirements.txt
phue
Pillow
requests